Use forDesignTime when default constructor has required parameters
In order to be showcased, the @Showcased widget class must have a default constructor with no required parameters without default values OR define a forDesignTime() factory constructor. The reason for this is that widgets are usually created with context specific constructor parameters. When rendering the preview of the widget, we don’t necessarily know the right values for these parameters. So, we let the user to provide them for us via this specially named constructor.

I'm using the same constructor IntelliJ uses for its live preview feature.

See https://github.com/flutter/flutter-intellij/wiki/Using-live-preview